Simple group_by function definition.
This is a simple group_by function that will take a container and sort the data into groups.
The group definition must be supplied by a lambda function.

The return type is an std::map<Key,DATA> where DATA is of the same type as the original container.
Note that the container that is returned from a key is not sorted in anyway.